+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 7.08.2018
 
+* [Rust] Code of Kutulu - Manhattan distance trait object (CodinGame)
 * [Rust] Code of Kutulu - Get coord in Entity trait (CodinGame)
 
 #### 6.08.2018